requirements.txt
python-dotenv==1.1.1
psycopg2-binary
supabase
